+++ src/components/index.jsx
@@ -5,5 +5,6 @@ export * from './Home';
 export * from './WhoWeAre';
 export * from './Teams';
 export * from './Partners';
+export * from  './Form';
