+++ src/pages/howToHelp/index.jsx
@@ -1,0 +1,1 @@
+export * from './HowToHelp'
