+++ src/pages/index.jsx
@@ -2,4 +2,6 @@ export * from './about';
 export * from './home';
 export * from './needHelp';
 export * from './partners';
-export * from './Team';
+export * from './team/Team';
+export * from './howToHelp';
+export * from './team';
